Function to generate a dashboard from a MAF file.

This function creates an HTML file containing the different figures and plots explaining the MAF dataset.

getMAFDashboard(
  MAFfilePath = NULL,
  plotList = NULL,
  outputFileName = "dashboard.html",
  outputFileTitle = "MAF Dash",
  masterRmdFile = NULL
)

Arguments

MAFfilePath
The path of the file containing the mutation information in the MAF format

plotList
A named list denoting which plots to draw. List elements can be: - boolean, if the name matches one of built-in plots: "summary_plot","burden","oncoplot","cooccurence","heatmap" - ggplot, ComplexHeatmap, or plotly object - file path to image (must be absolute path) The default value (if set to NULL) is 'list("summary_plot"=TRUE,"burden"=TRUE,"oncoplot"=TRUE,"cooccurence"=TRUE,"heatmap"=TRUE)' The order of the list determines the order of the tabs, and list names are used as tab titles.

outputFileName
The name of the output html dashboard file

outputFileTitle
The title of the output html dashboard

masterRmdFile
Alternate Rmd file. This argument is meant only for testing and development. file

Value

The dashboard html file

Author

Mayank Tondon, Ashish Jain

Examples

library(MAFDashRPackage)
#MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)